+++ AuthenticationService/Models/AccountPinStatusResponse.cs
@@ -3,6 +3,6 @@ namespace AuthenticationService.Models
     public class AccountPinStatusResponse
     {
         public bool IsEnabled { get; set; }
-        public double? UnlockedUntil { get; set; }
+        public long? UnlockedUntil { get; set; }
     }
 }
